// src/app/api/model/collectionPart.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CollectionPart { 
    adult?: boolean;
    backdrop_path?: string | null;
    id?: number;
    title?: string;
    original_language?: string;
    original_title?: string;
    overview?: string;
    poster_path?: string | null;
    media_type?: string;
    genre_ids?: Array<number>;
    popularity?: number;
    release_date?: string;
    video?: boolean;
    vote_average?: number;
    vote_count?: number;
}